chore: partially port `route` - interdependent types
