+++ packages/svelte-router/src/location.ts
