+++ packages/svelte-router/src/route.ts
@@ -1,12 +1,1662 @@
+import invariant from 'tiny-invariant'
+// import { useMatch } from './useMatch'
+// import { useLoaderDeps } from './useLoaderDeps'
+// import { useLoaderData } from './useLoaderData'
+// import { joinPaths, trimPathLeft } from './path'
+// import { useParams } from './useParams'
+// import { useSearch } from './useSearch'
+// import { notFound } from './not-found'
+// import { useNavigate } from './useNavigate'
+// import { rootRouteId } from './root'
+import type { RootRouteId } from './root'
+// import type { UseNavigateResult } from './useNavigate'
+// import type {
+// MakeRouteMatch,
+// MakeRouteMatchFromRoute,
+// MakeRouteMatchUnion,
+// RouteMatch,
+// } from './Matches'
+import type { ParsePathParams } from './link'
+import type { ParsedLocation } from './location'
+// import type { RouteById, RouteIds, RoutePaths } from './routeInfo'
+// import type { AnyRouter, RegisteredRouter, Router } from './router'
+import type { Assign, Constrain, Expand, NoInfer } from './utils'
 import type { Component } from 'svelte'
+// import type { BuildLocationFn, NavigateFn } from './RouterProvider'
+type BuildLocationFn = any
+type NavigateFn = any
+// import type { NotFoundError } from './not-found'
+// import type { LazyRoute } from './fileRoute'
 
 export type AnyPathParams = {}
 
+export type SearchSchemaInput = {
+  __TSearchSchemaInput__: 'TSearchSchemaInput'
+}
+
 export type AnySearchSchema = {}
+
+export type AnyContext = {}
+
+export interface RouteContext { }
+
+export type PreloadableObj = { preload?: () => Promise<void> }
+
+export type RoutePathOptions<TCustomId, TPath> =
+  | {
+    path: TPath
+  }
+  | {
+    id: TCustomId
+  }
+
+export interface StaticDataRouteOption { }
+
+export type RoutePathOptionsIntersection<TCustomId, TPath> = {
+  path: TPath
+  id: TCustomId
+}
+
+export type RouteOptions<
+  TParentRoute extends AnyRoute = AnyRoute,
+  TId extends string = string,
+  TCustomId extends string = string,
+  TFullPath extends string = string,
+  TPath extends string = string,
+  TSearchValidator = undefined,
+  TParams = AnyPathParams,
+  TLoaderDeps extends Record<string, any> = {},
+  TLoaderFn = undefined,
+  TRouterContext = {},
+  TRouteContextFn = AnyContext,
+  TBeforeLoadFn = AnyContext,
+> = BaseRouteOptions<
+  TParentRoute,
+  TId,
+  TCustomId,
+  TPath,
+  TSearchValidator,
+  TParams,
+  TLoaderDeps,
+  TLoaderFn,
+  TRouterContext,
+  TRouteContextFn,
+  TBeforeLoadFn
+> &
+  UpdatableRouteOptions<
+    NoInfer<TParentRoute>,
+    NoInfer<TCustomId>,
+    NoInfer<TFullPath>,
+    NoInfer<TParams>,
+    NoInfer<TSearchValidator>,
+    NoInfer<TLoaderFn>,
+    NoInfer<TLoaderDeps>,
+    NoInfer<TRouterContext>,
+    NoInfer<TRouteContextFn>,
+    NoInfer<TBeforeLoadFn>
+  >
+
+export type ParseSplatParams<TPath extends string> = TPath extends `${string}$`
+  ? '_splat'
+  : TPath extends `${string}$/${string}`
+  ? '_splat'
+  : never
+
+export interface SplatParams {
+  _splat?: string
+}
+
+export type ResolveParams<TPath extends string> =
+  ParseSplatParams<TPath> extends never
+  ? Record<ParsePathParams<TPath>, string>
+  : Record<ParsePathParams<TPath>, string> & SplatParams
+
+export type ParseParamsFn<TPath extends string, TParams> = (
+  rawParams: ResolveParams<TPath>,
+) => TParams extends Record<ParsePathParams<TPath>, any>
+  ? TParams
+  : Record<ParsePathParams<TPath>, any>
+
+export type StringifyParamsFn<TPath extends string, TParams> = (
+  params: TParams,
+) => ResolveParams<TPath>
+
+export type ParamsOptions<TPath extends string, TParams> = {
+  params?: {
+    parse?: ParseParamsFn<TPath, TParams>
+    stringify?: StringifyParamsFn<TPath, TParams>
+  }
+
+  /** 
+  @deprecated Use params.parse instead
+  */
+  parseParams?: ParseParamsFn<TPath, TParams>
+
+  /** 
+  @deprecated Use params.stringify instead
+  */
+  stringifyParams?: StringifyParamsFn<TPath, TParams>
+}
+
+export interface FullSearchSchemaOption<
+  in out TParentRoute extends AnyRoute,
+  in out TSearchValidator,
+> {
+  search: Expand<ResolveFullSearchSchema<TParentRoute, TSearchValidator>>
+}
+
+export type RouteContextFn<
+  in out TParentRoute extends AnyRoute,
+  in out TSearchValidator,
+  in out TParams,
+  in out TRouterContext,
+> = (
+  ctx: RouteContextOptions<
+    TParentRoute,
+    TSearchValidator,
+    TParams,
+    TRouterContext
+  >,
+) => any
+
+export type BeforeLoadFn<
+  in out TParentRoute extends AnyRoute,
+  in out TSearchValidator,
+  in out TParams,
+  in out TRouterContext,
+  in out TRouteContextFn,
+> = (
+  ctx: BeforeLoadContextOptions<
+    TParentRoute,
+    TSearchValidator,
+    TParams,
+    TRouterContext,
+    TRouteContextFn
+  >,
+) => any
+
+export type FileBaseRouteOptions<
+  TParentRoute extends AnyRoute = AnyRoute,
+  TId extends string = string,
+  TPath extends string = string,
+  TSearchValidator = undefined,
+  TParams = {},
+  TLoaderDeps extends Record<string, any> = {},
+  TLoaderFn = undefined,
+  TRouterContext = {},
+  TRouteContextFn = AnyContext,
+  TBeforeLoadFn = AnyContext,
+> = ParamsOptions<TPath, TParams> & {
+  validateSearch?: Constrain<
+    TSearchValidator,
+    AnySearchValidator,
+    DefaultSearchValidator
+  >
+
+  shouldReload?:
+  | boolean
+  // | ((
+  //   match: LoaderFnContext<
+  //     TParentRoute,
+  //     TId,
+  //     TParams,
+  //     TLoaderDeps,
+  //     TRouterContext,
+  //     TRouteContextFn,
+  //     TBeforeLoadFn
+  //   >,
+  // ) => any)
+
+  context?: Constrain<
+    TRouteContextFn,
+    (
+      ctx: RouteContextOptions<
+        TParentRoute,
+        TParams,
+        TRouterContext,
+        TLoaderDeps
+      >,
+    ) => any
+  >
+
+  // This async function is called before a route is loaded.
+  // If an error is thrown here, the route's loader will not be called.
+  // If thrown during a navigation, the navigation will be cancelled and the error will be passed to the `onError` function.
+  // If thrown during a preload event, the error will be logged to the console.
+  beforeLoad?: Constrain<
+    TBeforeLoadFn,
+    (
+      ctx: BeforeLoadContextOptions<
+        TParentRoute,
+        TSearchValidator,
+        TParams,
+        TRouterContext,
+        TRouteContextFn
+      >,
+    ) => any
+  >
+
+  loaderDeps?: (
+    opts: FullSearchSchemaOption<TParentRoute, TSearchValidator>,
+  ) => TLoaderDeps
+
+  // loader?: Constrain<
+  //   TLoaderFn,
+  //   (
+  //     ctx: LoaderFnContext<
+  //       TParentRoute,
+  //       TId,
+  //       TParams,
+  //       TLoaderDeps,
+  //       TRouterContext,
+  //       TRouteContextFn,
+  //       TBeforeLoadFn
+  //     >,
+  //   ) => any
+  // >
+}
+
+export type BaseRouteOptions<
+  TParentRoute extends AnyRoute = AnyRoute,
+  TId extends string = string,
+  TCustomId extends string = string,
+  TPath extends string = string,
+  TSearchValidator = undefined,
+  TParams = {},
+  TLoaderDeps extends Record<string, any> = {},
+  TLoaderFn = undefined,
+  TRouterContext = {},
+  TRouteContextFn = AnyContext,
+  TBeforeLoadFn = AnyContext,
+> = RoutePathOptions<TCustomId, TPath> &
+  FileBaseRouteOptions<
+    TParentRoute,
+    TId,
+    TPath,
+    TSearchValidator,
+    TParams,
+    TLoaderDeps,
+    TLoaderFn,
+    TRouterContext,
+    TRouteContextFn,
+    TBeforeLoadFn
+  > & {
+    getParentRoute: () => TParentRoute
+  }
+
+export interface ContextOptions<
+  in out TParentRoute extends AnyRoute,
+  in out TParams,
+> {
+  abortController: AbortController
+  preload: boolean
+  params: Expand<ResolveAllParamsFromParent<TParentRoute, TParams>>
+  location: ParsedLocation
+  /**
+   * @deprecated Use `throw redirect({ to: '/somewhere' })` instead
+   **/
+  navigate: NavigateFn
+  buildLocation: BuildLocationFn
+  cause: 'preload' | 'enter' | 'stay'
+  // matches: Array<MakeRouteMatchUnion>
+}
+
+export interface RouteContextOptions<
+  in out TParentRoute extends AnyRoute,
+  in out TParams,
+  in out TRouterContext,
+  in out TLoaderDeps,
+> extends ContextOptions<TParentRoute, TParams> {
+  deps: TLoaderDeps
+  context: Expand<RouteContextParameter<TParentRoute, TRouterContext>>
+}
+
+export interface BeforeLoadContextOptions<
+  in out TParentRoute extends AnyRoute,
+  in out TSearchValidator,
+  in out TParams,
+  in out TRouterContext,
+  in out TRouteContextFn,
+> extends ContextOptions<TParentRoute, TParams>,
+  FullSearchSchemaOption<TParentRoute, TSearchValidator> {
+  context: Expand<
+    BeforeLoadContextParameter<TParentRoute, TRouterContext, TRouteContextFn>
+  >
+}
+
+export interface UpdatableRouteOptions<
+  in out TParentRoute extends AnyRoute,
+  in out TRouteId,
+  in out TFullPath,
+  in out TParams,
+  in out TSearchValidator,
+  in out TLoaderFn,
+  in out TLoaderDeps,
+  in out TRouterContext,
+  in out TRouteContextFn,
+  in out TBeforeLoadFn,
+> extends UpdatableStaticRouteOption {
+  // If true, this route will be matched as case-sensitive
+  caseSensitive?: boolean
+  // If true, this route will be forcefully wrapped in a suspense boundary
+  wrapInSuspense?: boolean
+  // The content to be rendered when the route is matched. If no component is provided, defaults to `<Outlet />`
+  component?: RouteComponent
+  errorComponent?: false | null | ErrorRouteComponent
+  notFoundComponent?: NotFoundRouteComponent
+  pendingComponent?: RouteComponent
+  pendingMs?: number
+  pendingMinMs?: number
+  staleTime?: number
+  gcTime?: number
+  preload?: boolean
+  preloadStaleTime?: number
+  preloadGcTime?: number
+  search?: {
+    middlewares?: Array<
+      SearchMiddleware<
+        ResolveFullSearchSchemaInput<TParentRoute, TSearchValidator>
+      >
+    >
+  }
+  /** 
+  @deprecated Use search.middlewares instead
+  */
+  preSearchFilters?: Array<
+    SearchFilter<ResolveFullSearchSchema<TParentRoute, TSearchValidator>>
+  >
+  /** 
+  @deprecated Use search.middlewares instead
+  */
+  postSearchFilters?: Array<
+    SearchFilter<ResolveFullSearchSchema<TParentRoute, TSearchValidator>>
+  >
+  onCatch?: (error: Error, errorInfo: any) => void
+  onError?: (err: any) => void
+  // These functions are called as route matches are loaded, stick around and leave the active
+  // matches
+  // onEnter?: (
+  //   match: RouteMatch<
+  //     TRouteId,
+  //     TFullPath,
+  //     ResolveAllParamsFromParent<TParentRoute, TParams>,
+  //     ResolveFullSearchSchema<TParentRoute, TSearchValidator>,
+  //     ResolveLoaderData<TLoaderFn>,
+  //     ResolveAllContext<
+  //       TParentRoute,
+  //       TRouterContext,
+  //       TRouteContextFn,
+  //       TBeforeLoadFn
+  //     >,
+  //     TLoaderDeps
+  //   >,
+  // ) => void
+  // onStay?: (
+  //   match: RouteMatch<
+  //     TRouteId,
+  //     TFullPath,
+  //     ResolveAllParamsFromParent<TParentRoute, TParams>,
+  //     ResolveFullSearchSchema<TParentRoute, TSearchValidator>,
+  //     ResolveLoaderData<TLoaderFn>,
+  //     ResolveAllContext<
+  //       TParentRoute,
+  //       TRouterContext,
+  //       TRouteContextFn,
+  //       TBeforeLoadFn
+  //     >,
+  //     TLoaderDeps
+  //   >,
+  // ) => void
+  // onLeave?: (
+  //   match: RouteMatch<
+  //     TRouteId,
+  //     TFullPath,
+  //     ResolveAllParamsFromParent<TParentRoute, TParams>,
+  //     ResolveFullSearchSchema<TParentRoute, TSearchValidator>,
+  //     ResolveLoaderData<TLoaderFn>,
+  //     ResolveAllContext<
+  //       TParentRoute,
+  //       TRouterContext,
+  //       TRouteContextFn,
+  //       TBeforeLoadFn
+  //     >,
+  //     TLoaderDeps
+  //   >,
+  // ) => void
+  // meta?: (ctx: {
+  //   matches: Array<
+  //     RouteMatch<
+  //       TRouteId,
+  //       TFullPath,
+  //       ResolveAllParamsFromParent<TParentRoute, TParams>,
+  //       ResolveFullSearchSchema<TParentRoute, TSearchValidator>,
+  //       ResolveLoaderData<TLoaderFn>,
+  //       ResolveAllContext<
+  //         TParentRoute,
+  //         TRouterContext,
+  //         TRouteContextFn,
+  //         TBeforeLoadFn
+  //       >,
+  //       TLoaderDeps
+  //     >
+  //   >
+  //   match: RouteMatch<
+  //     TRouteId,
+  //     TFullPath,
+  //     ResolveAllParamsFromParent<TParentRoute, TParams>,
+  //     ResolveFullSearchSchema<TParentRoute, TSearchValidator>,
+  //     ResolveLoaderData<TLoaderFn>,
+  //     ResolveAllContext<
+  //       TParentRoute,
+  //       TRouterContext,
+  //       TRouteContextFn,
+  //       TBeforeLoadFn
+  //     >,
+  //     TLoaderDeps
+  //   >
+  //   params: ResolveAllParamsFromParent<TParentRoute, TParams>
+  //   loaderData: ResolveLoaderData<TLoaderFn>
+  // }) => Array<React.JSX.IntrinsicElements['meta']>
+  // links?: () => Array<React.JSX.IntrinsicElements['link']>
+  // scripts?: () => Array<React.JSX.IntrinsicElements['script']>
+  headers?: (ctx: {
+    loaderData: ResolveLoaderData<TLoaderFn>
+  }) => Record<string, string>
+  ssr?: boolean
+}
+
+interface RequiredStaticDataRouteOption {
+  staticData: StaticDataRouteOption
+}
+
+interface OptionalStaticDataRouteOption {
+  staticData?: StaticDataRouteOption
+}
+
+export type UpdatableStaticRouteOption = {} extends StaticDataRouteOption
+  ? OptionalStaticDataRouteOption
+  : RequiredStaticDataRouteOption
+
+export type MetaDescriptor =
+  | { charSet: 'utf-8' }
+  | { title: string }
+  | { name: string; content: string }
+  | { property: string; content: string }
+  | { httpEquiv: string; content: string }
+  | { 'script:ld+json': LdJsonObject }
+  | { tagName: 'meta' | 'link';[name: string]: string }
+  | Record<string, unknown>
+
+type LdJsonObject = { [Key in string]: LdJsonValue } & {
+  [Key in string]?: LdJsonValue | undefined
+}
+type LdJsonArray = Array<LdJsonValue> | ReadonlyArray<LdJsonValue>
+type LdJsonPrimitive = string | number | boolean | null
+type LdJsonValue = LdJsonPrimitive | LdJsonObject | LdJsonArray
+
+export type RouteLinkEntry = {}
+
+export interface SearchValidatorObj<TInput, TOutput> {
+  parse: SearchValidatorFn<TInput, TOutput>
+}
+
+export type AnySearchValidatorObj = SearchValidatorObj<any, any>
+
+export interface SearchValidatorAdapter<TInput, TOutput> {
+  types: {
+    input: TInput
+    output: TOutput
+  }
+  parse: (input: unknown) => TOutput
+}
+
+export type AnySearchValidatorAdapter = SearchValidatorAdapter<any, any>
+
+export interface StandardSchemaValidator<TInput, TOutput> {
+  readonly '~types'?: StandardSchemaValidatorTypes<TInput, TOutput> | undefined
+  readonly '~validate': AnyStandardSchemaValidate
+}
+
+export type AnyStandardSchemaValidator = StandardSchemaValidator<any, any>
+
+export interface StandardSchemaValidatorTypes<TInput, TOutput> {
+  readonly input: TInput
+  readonly output: TOutput
+}
+
+export interface AnyStandardSchemaValidateSuccess {
+  readonly value: any
+}
+
+export interface AnyStandardSchemaValidateFailure {
+  readonly issues: ReadonlyArray<AnyStandardSchemaValidateIssue>
+}
+
+export interface AnyStandardSchemaValidateIssue {
+  readonly message: string
+}
+
+export interface AnyStandardSchemaValidateInput {
+  readonly value: any
+}
+
+export type AnyStandardSchemaValidate = (
+  input: AnyStandardSchemaValidateInput,
+) => AnyStandardSchemaValidateSuccess | AnyStandardSchemaValidateFailure
+
+export type AnySearchValidatorFn = SearchValidatorFn<any, any>
+
+export type SearchValidatorFn<TInput, TOutput> = (input: TInput) => TOutput
+
+export type SearchValidator<TInput, TOutput> =
+  | SearchValidatorObj<TInput, TOutput>
+  | SearchValidatorFn<TInput, TOutput>
+  | SearchValidatorAdapter<TInput, TOutput>
+  | StandardSchemaValidator<TInput, TOutput>
+  | undefined
+
+export type AnySearchValidator = SearchValidator<any, any>
+
+export type DefaultSearchValidator = SearchValidator<
+  Record<string, unknown>,
+  AnySearchSchema
+>
+
+// export type RouteLoaderFn<
+//   in out TParentRoute extends AnyRoute = AnyRoute,
+//   in out TId extends string = string,
+//   in out TParams = {},
+//   in out TLoaderDeps = {},
+//   in out TRouterContext = {},
+//   in out TRouteContextFn = AnyContext,
+//   in out TBeforeLoadFn = AnyContext,
+// > = (
+//   match: LoaderFnContext<
+//     TParentRoute,
+//     TId,
+//     TParams,
+//     TLoaderDeps,
+//     TRouterContext,
+//     TRouteContextFn,
+//     TBeforeLoadFn
+//   >,
+// ) => any
+
+// export interface LoaderFnContext<
+//   in out TParentRoute extends AnyRoute = AnyRoute,
+//   in out TId extends string = string,
+//   in out TParams = {},
+//   in out TLoaderDeps = {},
+//   in out TRouterContext = {},
+//   in out TRouteContextFn = AnyContext,
+//   in out TBeforeLoadFn = AnyContext,
+// > {
+//   abortController: AbortController
+//   preload: boolean
+//   params: Expand<ResolveAllParamsFromParent<TParentRoute, TParams>>
+//   deps: TLoaderDeps
+//   context: Expand<
+//     ResolveAllContext<
+//       TParentRoute,
+//       TRouterContext,
+//       TRouteContextFn,
+//       TBeforeLoadFn
+//     >
+//   >
+//   location: ParsedLocation // Do not supply search schema here so as to demotivate people from trying to shortcut loaderDeps
+//   /**
+//    * @deprecated Use `throw redirect({ to: '/somewhere' })` instead
+//    **/
+//   navigate: (opts: NavigateOptions<AnyRouter>) => Promise<void>
+//   // root route does not have a parent match
+//   parentMatchPromise: TId extends RootRouteId
+//   ? never
+//   : Promise<MakeRouteMatchFromRoute<TParentRoute>>
+//   cause: 'preload' | 'enter' | 'stay'
+//   route: Route
+// }
+
+export type SearchFilter<TInput, TResult = TInput> = (prev: TInput) => TResult
 
 export type SearchMiddlewareContext<TSearchSchema> = {
   search: TSearchSchema
   next: (newSearch: TSearchSchema) => TSearchSchema
+}
+
+export type SearchMiddleware<TSearchSchema> = (
+  ctx: SearchMiddlewareContext<TSearchSchema>,
+) => TSearchSchema
+
+export type ResolveId<
+  TParentRoute,
+  TCustomId extends string,
+  TPath extends string,
+> = TParentRoute extends { id: infer TParentId extends string }
+  ? RoutePrefix<TParentId, string extends TCustomId ? TPath : TCustomId>
+  : RootRouteId
+
+export type InferFullSearchSchema<TRoute> = TRoute extends {
+  types: {
+    fullSearchSchema: infer TFullSearchSchema
+  }
+}
+  ? TFullSearchSchema
+  : {}
+
+export type InferFullSearchSchemaInput<TRoute> = TRoute extends {
+  types: {
+    fullSearchSchemaInput: infer TFullSearchSchemaInput
+  }
+}
+  ? TFullSearchSchemaInput
+  : {}
+
+export type InferAllParams<TRoute> = TRoute extends {
+  types: {
+    allParams: infer TAllParams
+  }
+}
+  ? TAllParams
+  : {}
+
+export type InferAllContext<TRoute> = unknown extends TRoute
+  ? TRoute
+  : TRoute extends {
+    types: {
+      allContext: infer TAllContext
+    }
+  }
+  ? TAllContext
+  : {}
+
+export type ResolveSearchSchemaFnInput<TSearchValidator> =
+  TSearchValidator extends (input: infer TSearchSchemaInput) => any
+  ? TSearchSchemaInput extends SearchSchemaInput
+  ? Omit<TSearchSchemaInput, keyof SearchSchemaInput>
+  : ResolveSearchSchemaFn<TSearchValidator>
+  : AnySearchSchema
+
+export type ResolveSearchSchemaInput<TSearchValidator> =
+  TSearchValidator extends AnyStandardSchemaValidator
+  ? NonNullable<TSearchValidator['~types']>['input']
+  : TSearchValidator extends AnySearchValidatorAdapter
+  ? TSearchValidator['types']['input']
+  : TSearchValidator extends AnySearchValidatorObj
+  ? ResolveSearchSchemaFnInput<TSearchValidator['parse']>
+  : ResolveSearchSchemaFnInput<TSearchValidator>
+
+export type ResolveSearchSchemaFn<TSearchValidator> = TSearchValidator extends (
+  ...args: any
+) => infer TSearchSchema
+  ? TSearchSchema
+  : AnySearchSchema
+
+export type ResolveSearchSchema<TSearchValidator> =
+  unknown extends TSearchValidator
+  ? TSearchValidator
+  : TSearchValidator extends AnyStandardSchemaValidator
+  ? NonNullable<TSearchValidator['~types']>['output']
+  : TSearchValidator extends AnySearchValidatorAdapter
+  ? TSearchValidator['types']['output']
+  : TSearchValidator extends AnySearchValidatorObj
+  ? ResolveSearchSchemaFn<TSearchValidator['parse']>
+  : ResolveSearchSchemaFn<TSearchValidator>
+
+export type ResolveFullSearchSchema<
+  TParentRoute extends AnyRoute,
+  TSearchValidator,
+> = unknown extends TParentRoute
+  ? ResolveSearchSchema<TSearchValidator>
+  : Assign<
+    InferFullSearchSchema<TParentRoute>,
+    ResolveSearchSchema<TSearchValidator>
+  >
+
+export type ResolveFullSearchSchemaInput<
+  TParentRoute extends AnyRoute,
+  TSearchValidator,
+> = Assign<
+  InferFullSearchSchemaInput<TParentRoute>,
+  ResolveSearchSchemaInput<TSearchValidator>
+>
+
+export type LooseReturnType<T> = T extends (
+  ...args: Array<any>
+) => infer TReturn
+  ? TReturn
+  : never
+
+export type LooseAsyncReturnType<T> = T extends (
+  ...args: Array<any>
+) => infer TReturn
+  ? TReturn extends Promise<infer TReturn>
+  ? TReturn
+  : TReturn
+  : never
+
+export type ContextReturnType<TContextFn> = unknown extends TContextFn
+  ? TContextFn
+  : LooseReturnType<TContextFn> extends never
+  ? AnyContext
+  : LooseReturnType<TContextFn>
+
+export type ContextAsyncReturnType<TContextFn> = unknown extends TContextFn
+  ? TContextFn
+  : LooseAsyncReturnType<TContextFn> extends never
+  ? AnyContext
+  : LooseAsyncReturnType<TContextFn>
+
+export type RouteContextParameter<
+  TParentRoute extends AnyRoute,
+  TRouterContext,
+> = unknown extends TParentRoute
+  ? TRouterContext
+  : Assign<TRouterContext, InferAllContext<TParentRoute>>
+
+export type ResolveRouteContext<TRouteContextFn, TBeforeLoadFn> = Assign<
+  ContextReturnType<TRouteContextFn>,
+  ContextAsyncReturnType<TBeforeLoadFn>
+>
+export type BeforeLoadContextParameter<
+  TParentRoute extends AnyRoute,
+  TRouterContext,
+  TRouteContextFn,
+> = Assign<
+  RouteContextParameter<TParentRoute, TRouterContext>,
+  ContextReturnType<TRouteContextFn>
+>
+
+export type ResolveAllContext<
+  TParentRoute extends AnyRoute,
+  TRouterContext,
+  TRouteContextFn,
+  TBeforeLoadFn,
+> = Assign<
+  BeforeLoadContextParameter<TParentRoute, TRouterContext, TRouteContextFn>,
+  ContextAsyncReturnType<TBeforeLoadFn>
+>
+
+export type ResolveLoaderData<TLoaderFn> = unknown extends TLoaderFn
+  ? TLoaderFn
+  : LooseAsyncReturnType<TLoaderFn> extends never
+  ? {}
+  : LooseAsyncReturnType<TLoaderFn>
+
+export interface AnyRoute
+  extends Route<
+    any,
+    any,
+    any,
+    any,
+    any,
+    any,
+    any,
+    any,
+    any,
+    any,
+    any,
+    any,
+    any
+  > { }
+
+export type AnyRouteWithContext<TContext> = AnyRoute & {
+  types: { allContext: TContext }
+}
+
+export type ResolveAllParamsFromParent<
+  TParentRoute extends AnyRoute,
+  TParams,
+> = Assign<InferAllParams<TParentRoute>, TParams>
+
+export type RouteConstraints = {
+  TParentRoute: AnyRoute
+  TPath: string
+  TFullPath: string
+  TCustomId: string
+  TId: string
+  TSearchSchema: AnySearchSchema
+  TFullSearchSchema: AnySearchSchema
+  TParams: Record<string, any>
+  TAllParams: Record<string, any>
+  TParentContext: AnyContext
+  TRouteContext: RouteContext
+  TAllContext: AnyContext
+  TRouterContext: AnyContext
+  TChildren: unknown
+  TRouteTree: AnyRoute
+}
+
+// export type RouteTypesById<TRouter extends AnyRouter, TId> = RouteById<
+//   TRouter['routeTree'],
+//   TId
+// >['types']
+
+// export function getRouteApi<TId, TRouter extends AnyRouter = RegisteredRouter>(
+//   id: Constrain<TId, RouteIds<TRouter['routeTree']>>,
+// ) {
+//   return new RouteApi<TId, TRouter>({ id })
+// }
+
+// export class RouteApi<TId, TRouter extends AnyRouter = RegisteredRouter> {
+//   id: TId
+
+//   /**
+//    * @deprecated Use the `getRouteApi` function instead.
+//    */
+//   constructor({ id }: { id: TId }) {
+//     this.id = id as any
+//   }
+
+// useMatch = <
+//   TRouteTree extends AnyRoute = TRouter['routeTree'],
+//   TRouteMatch = MakeRouteMatch<TRouteTree, TId>,
+//   TSelected = TRouteMatch,
+// >(opts?: {
+//   select?: (match: TRouteMatch) => TSelected
+// }): TSelected => {
+//   return useMatch({ select: opts?.select, from: this.id })
+// }
+
+// useRouteContext = <
+//   TSelected = Expand<RouteTypesById<TRouter, TId>['allContext']>,
+// >(opts?: {
+//   select?: (
+//     s: Expand<RouteTypesById<TRouter, TId>['allContext']>,
+//   ) => TSelected
+// }): TSelected => {
+//   return useMatch({
+//     from: this.id,
+//     select: (d: any) => (opts?.select ? opts.select(d.context) : d.context),
+//   })
+// }
+
+// useSearch = <
+//   TSelected = Expand<RouteTypesById<TRouter, TId>['fullSearchSchema']>,
+// >(opts?: {
+//   select?: (
+//     s: Expand<RouteTypesById<TRouter, TId>['fullSearchSchema']>,
+//   ) => TSelected
+// }): TSelected => {
+//   return useSearch({ ...opts, from: this.id })
+// }
+
+// useParams = <
+//   TSelected = Expand<RouteTypesById<TRouter, TId>['allParams']>,
+// >(opts?: {
+//   select?: (s: Expand<RouteTypesById<TRouter, TId>['allParams']>) => TSelected
+// }): TSelected => {
+//   return useParams({ ...opts, from: this.id })
+// }
+
+// useLoaderDeps = <
+//   TSelected = RouteTypesById<TRouter, TId>['loaderDeps'],
+// >(opts?: {
+//   select?: (s: RouteTypesById<TRouter, TId>['loaderDeps']) => TSelected
+// }): TSelected => {
+//   return useLoaderDeps({ ...opts, from: this.id, strict: false } as any)
+// }
+
+// useLoaderData = <
+//   TSelected = RouteTypesById<TRouter, TId>['loaderData'],
+// >(opts?: {
+//   select?: (s: RouteTypesById<TRouter, TId>['loaderData']) => TSelected
+// }): TSelected => {
+//   return useLoaderData({ ...opts, from: this.id, strict: false } as any)
+// }
+
+// useNavigate = (): UseNavigateResult<
+//   RouteTypesById<TRouter, TId>['fullPath']
+// > => {
+//   return useNavigate({ from: this.id as string })
+// }
+
+//   notFound = (opts?: NotFoundError) => {
+//     return notFound({ routeId: this.id as string, ...opts })
+//   }
+// }
+
+export class Route<
+  in out TParentRoute extends RouteConstraints['TParentRoute'] = AnyRoute,
+  in out TPath extends RouteConstraints['TPath'] = '/',
+  in out TFullPath extends RouteConstraints['TFullPath'] = ResolveFullPath<
+    TParentRoute,
+    TPath
+  >,
+  in out TCustomId extends RouteConstraints['TCustomId'] = string,
+  in out TId extends RouteConstraints['TId'] = ResolveId<
+    TParentRoute,
+    TCustomId,
+    TPath
+  >,
+  in out TSearchValidator = undefined,
+  in out TParams = ResolveParams<TPath>,
+  in out TRouterContext = AnyContext,
+  in out TRouteContextFn = AnyContext,
+  in out TBeforeLoadFn = AnyContext,
+  in out TLoaderDeps extends Record<string, any> = {},
+  in out TLoaderFn = undefined,
+  in out TChildren = unknown,
+> {
+  isRoot: TParentRoute extends Route<any> ? true : false
+  options: RouteOptions<
+    TParentRoute,
+    TId,
+    TCustomId,
+    TFullPath,
+    TPath,
+    TSearchValidator,
+    TParams,
+    TLoaderDeps,
+    TLoaderFn,
+    TRouterContext,
+    TRouteContextFn,
+    TBeforeLoadFn
+  >
+
+  // The following properties are set up in this.init()
+  parentRoute!: TParentRoute
+  private _id!: TId
+  private _path!: TPath
+  private _fullPath!: TFullPath
+  private _to!: TrimPathRight<TFullPath>
+  private _ssr!: boolean
+
+  public get to() {
+    /* invariant(
+      this._to,
+      `trying to access property 'to' on a route which is not initialized yet. Route properties are only available after 'createRouter' completed.`,
+    )*/
+    return this._to
+  }
+
+  public get id() {
+    /* invariant(
+      this._id,
+      `trying to access property 'id' on a route which is not initialized yet. Route properties are only available after 'createRouter' completed.`,
+    )*/
+    return this._id
+  }
+
+  public get path() {
+    /* invariant(
+      // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
+      this.isRoot || this._id || this._path,
+      `trying to access property 'path' on a route which is not initialized yet. Route properties are only available after 'createRouter' completed.`,
+    )*/
+    return this._path
+  }
+
+  public get fullPath() {
+    /* invariant(
+      this._fullPath,
+      `trying to access property 'fullPath' on a route which is not initialized yet. Route properties are only available after 'createRouter' completed.`,
+    )*/
+    return this._fullPath
+  }
+
+  public get ssr() {
+    return this._ssr
+  }
+
+  // Optional
+  children?: TChildren
+  originalIndex?: number
+  // router?: AnyRouter
+  rank!: number
+  // lazyFn?: () => Promise<LazyRoute<any>>
+  _lazyPromise?: Promise<void>
+  _componentsPromise?: Promise<Array<void>>
+
+  /**
+   * @deprecated Use the `createRoute` function instead.
+   */
+  constructor(
+    options?: RouteOptions<
+      TParentRoute,
+      TId,
+      TCustomId,
+      TFullPath,
+      TPath,
+      TSearchValidator,
+      TParams,
+      TLoaderDeps,
+      TLoaderFn,
+      TRouterContext,
+      TRouteContextFn,
+      TBeforeLoadFn
+    >,
+  ) {
+    this.options = (options as any) || {}
+
+    this.isRoot = !options?.getParentRoute as any
+    invariant(
+      !((options as any)?.id && (options as any)?.path),
+      `Route cannot have both an 'id' and a 'path' option.`,
+    )
+      ; (this as any).$$typeof = Symbol.for('react.memo')
+  }
+
+  types!: {
+    parentRoute: TParentRoute
+    path: TPath
+    to: TrimPathRight<TFullPath>
+    fullPath: TFullPath
+    customId: TCustomId
+    id: TId
+    searchSchema: ResolveSearchSchema<TSearchValidator>
+    searchSchemaInput: ResolveSearchSchemaInput<TSearchValidator>
+    searchValidator: TSearchValidator
+    fullSearchSchema: ResolveFullSearchSchema<TParentRoute, TSearchValidator>
+    fullSearchSchemaInput: ResolveFullSearchSchemaInput<
+      TParentRoute,
+      TSearchValidator
+    >
+    params: TParams
+    allParams: ResolveAllParamsFromParent<TParentRoute, TParams>
+    routerContext: TRouterContext
+    routeContext: ResolveRouteContext<TRouteContextFn, TBeforeLoadFn>
+    routeContextFn: TRouteContextFn
+    beforeLoadFn: TBeforeLoadFn
+    allContext: ResolveAllContext<
+      TParentRoute,
+      TRouterContext,
+      TRouteContextFn,
+      TBeforeLoadFn
+    >
+    children: TChildren
+    loaderData: ResolveLoaderData<TLoaderFn>
+    loaderDeps: TLoaderDeps
+  }
+
+  // init = (opts: { originalIndex: number; defaultSsr?: boolean }): void => {
+  //   this.originalIndex = opts.originalIndex
+
+  //   const options = this.options as
+  //     | (RouteOptions<
+  //       TParentRoute,
+  //       TId,
+  //       TCustomId,
+  //       TFullPath,
+  //       TPath,
+  //       TSearchValidator,
+  //       TParams,
+  //       TLoaderDeps,
+  //       TLoaderFn,
+  //       TRouterContext,
+  //       TRouteContextFn,
+  //       TBeforeLoadFn
+  //     > &
+  //       RoutePathOptionsIntersection<TCustomId, TPath>)
+  //     | undefined
+
+  //   const isRoot = !options?.path && !options?.id
+
+  //   // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
+  //   this.parentRoute = this.options.getParentRoute?.()
+
+  //   if (isRoot) {
+  //     this._path = rootRouteId as TPath
+  //   } else {
+  //     invariant(
+  //       this.parentRoute,
+  //       `Child Route instances must pass a 'getParentRoute: () => ParentRoute' option that returns a Route instance.`,
+  //     )
+  //   }
+
+  //   let path: undefined | string = isRoot ? rootRouteId : options.path
+
+  //   // If the path is anything other than an index path, trim it up
+  //   if (path && path !== '/') {
+  //     path = trimPathLeft(path)
+  //   }
+
+  //   const customId = options?.id || path
+
+  //   // Strip the parentId prefix from the first level of children
+  //   let id = isRoot
+  //     ? rootRouteId
+  //     : joinPaths([
+  //       this.parentRoute.id === rootRouteId ? '' : this.parentRoute.id,
+  //       customId,
+  //     ])
+
+  //   if (path === rootRouteId) {
+  //     path = '/'
+  //   }
+
+  //   if (id !== rootRouteId) {
+  //     id = joinPaths(['/', id])
+  //   }
+
+  //   const fullPath =
+  //     id === rootRouteId ? '/' : joinPaths([this.parentRoute.fullPath, path])
+
+  //   this._path = path as TPath
+  //   this._id = id as TId
+  //   // this.customId = customId as TCustomId
+  //   this._fullPath = fullPath as TFullPath
+  //   this._to = fullPath as TrimPathRight<TFullPath>
+  //   this._ssr = options?.ssr ?? opts.defaultSsr ?? true
+  // }
+
+  addChildren<
+    const TNewChildren extends
+    | Record<string, AnyRoute>
+    | ReadonlyArray<AnyRoute>,
+  >(
+    children: TNewChildren,
+  ): Route<
+    TParentRoute,
+    TPath,
+    TFullPath,
+    TCustomId,
+    TId,
+    TSearchValidator,
+    TParams,
+    TRouterContext,
+    TRouteContextFn,
+    TBeforeLoadFn,
+    TLoaderDeps,
+    TLoaderFn,
+    TNewChildren
+  > {
+    return this._addFileChildren(children)
+  }
+
+  _addFileChildren<const TNewChildren>(
+    children: TNewChildren,
+  ): Route<
+    TParentRoute,
+    TPath,
+    TFullPath,
+    TCustomId,
+    TId,
+    TSearchValidator,
+    TParams,
+    TRouterContext,
+    TRouteContextFn,
+    TBeforeLoadFn,
+    TLoaderDeps,
+    TLoaderFn,
+    TNewChildren
+  > {
+    if (Array.isArray(children)) {
+      this.children = children as TChildren
+    }
+
+    if (typeof children === 'object' && children !== null) {
+      this.children = Object.values(children) as TChildren
+    }
+
+    return this as unknown as Route<
+      TParentRoute,
+      TPath,
+      TFullPath,
+      TCustomId,
+      TId,
+      TSearchValidator,
+      TParams,
+      TRouterContext,
+      TRouteContextFn,
+      TBeforeLoadFn,
+      TLoaderDeps,
+      TLoaderFn,
+      TNewChildren
+    >
+  }
+
+  // updateLoader = <TNewLoaderFn>(options: {
+  //   loader: Constrain<
+  //     TNewLoaderFn,
+  //     RouteLoaderFn<
+  //       TParentRoute,
+  //       TCustomId,
+  //       TParams,
+  //       TLoaderDeps,
+  //       TRouterContext,
+  //       TRouteContextFn,
+  //       TBeforeLoadFn
+  //     >
+  //   >
+  // }) => {
+  //   Object.assign(this.options, options)
+  //   return this as unknown as Route<
+  //     TParentRoute,
+  //     TPath,
+  //     TFullPath,
+  //     TCustomId,
+  //     TId,
+  //     TSearchValidator,
+  //     TParams,
+  //     TRouterContext,
+  //     TRouteContextFn,
+  //     TBeforeLoadFn,
+  //     TLoaderDeps,
+  //     TNewLoaderFn,
+  //     TChildren
+  //   >
+  // }
+
+  // update = (
+  //   options: UpdatableRouteOptions<
+  //     TParentRoute,
+  //     TCustomId,
+  //     TFullPath,
+  //     TParams,
+  //     TSearchValidator,
+  //     TLoaderFn,
+  //     TLoaderDeps,
+  //     TRouterContext,
+  //     TRouteContextFn,
+  //     TBeforeLoadFn
+  //   >,
+  // ): this => {
+  //   Object.assign(this.options, options)
+  //   return this
+  // }
+
+  // lazy = (lazyFn: () => Promise<LazyRoute<any>>): this => {
+  //   this.lazyFn = lazyFn
+  //   return this
+  // }
+
+  // useMatch = <
+  //   TRouter extends AnyRouter = RegisteredRouter,
+  //   TRouteTree extends AnyRoute = TRouter['routeTree'],
+  //   TRouteMatch = MakeRouteMatch<TRouteTree, TId>,
+  //   TSelected = TRouteMatch,
+  // >(opts?: {
+  //   select?: (match: TRouteMatch) => TSelected
+  // }): TSelected => {
+  //   return useMatch({ ...opts, from: this.id })
+  // }
+
+  // useRouteContext = <
+  //   TSelected = Expand<
+  //     ResolveAllContext<
+  //       TParentRoute,
+  //       TRouterContext,
+  //       TRouteContextFn,
+  //       TBeforeLoadFn
+  //     >
+  //   >,
+  // >(opts?: {
+  //   select?: (
+  //     search: Expand<
+  //       ResolveAllContext<
+  //         TParentRoute,
+  //         TRouterContext,
+  //         TRouteContextFn,
+  //         TBeforeLoadFn
+  //       >
+  //     >,
+  //   ) => TSelected
+  // }): TSelected => {
+  //   return useMatch({
+  //     ...opts,
+  //     from: this.id,
+  //     select: (d: any) => (opts?.select ? opts.select(d.context) : d.context),
+  //   })
+  // }
+
+  // useSearch = <
+  //   TSelected = Expand<ResolveFullSearchSchema<TParentRoute, TSearchValidator>>,
+  // >(opts?: {
+  //   select?: (
+  //     search: Expand<ResolveFullSearchSchema<TParentRoute, TSearchValidator>>,
+  //   ) => TSelected
+  // }): TSelected => {
+  //   return useSearch({ ...opts, from: this.id })
+  // }
+
+  // useParams = <
+  //   TSelected = Expand<ResolveAllParamsFromParent<TParentRoute, TParams>>,
+  // >(opts?: {
+  //   select?: (
+  //     search: Expand<ResolveAllParamsFromParent<TParentRoute, TParams>>,
+  //   ) => TSelected
+  // }): TSelected => {
+  //   return useParams({ ...opts, from: this.id })
+  // }
+
+  // useLoaderDeps = <TSelected = TLoaderDeps>(opts?: {
+  //   select?: (s: TLoaderDeps) => TSelected
+  // }): TSelected => {
+  //   return useLoaderDeps({ ...opts, from: this.id } as any)
+  // }
+
+  // useLoaderData = <TSelected = ResolveLoaderData<TLoaderFn>>(opts?: {
+  //   select?: (search: ResolveLoaderData<TLoaderFn>) => TSelected
+  // }): TSelected => {
+  //   return useLoaderData({ ...opts, from: this.id } as any)
+  // }
+
+  // useNavigate = (): UseNavigateResult<TFullPath> => {
+  //   return useNavigate({ from: this.id })
+  // }
+}
+
+export function createRoute<
+  TParentRoute extends RouteConstraints['TParentRoute'] = AnyRoute,
+  TPath extends RouteConstraints['TPath'] = '/',
+  TFullPath extends RouteConstraints['TFullPath'] = ResolveFullPath<
+    TParentRoute,
+    TPath
+  >,
+  TCustomId extends RouteConstraints['TCustomId'] = string,
+  TId extends RouteConstraints['TId'] = ResolveId<
+    TParentRoute,
+    TCustomId,
+    TPath
+  >,
+  TSearchValidator = undefined,
+  TParams = ResolveParams<TPath>,
+  TRouteContextFn = AnyContext,
+  TBeforeLoadFn = AnyContext,
+  TLoaderDeps extends Record<string, any> = {},
+  TLoaderFn = undefined,
+  TChildren = unknown,
+>(
+  options: RouteOptions<
+    TParentRoute,
+    TId,
+    TCustomId,
+    TFullPath,
+    TPath,
+    TSearchValidator,
+    TParams,
+    TLoaderDeps,
+    TLoaderFn,
+    AnyContext,
+    TRouteContextFn,
+    TBeforeLoadFn
+  >,
+) {
+  return new Route<
+    TParentRoute,
+    TPath,
+    TFullPath,
+    TCustomId,
+    TId,
+    TSearchValidator,
+    TParams,
+    AnyContext,
+    TRouteContextFn,
+    TBeforeLoadFn,
+    TLoaderDeps,
+    TLoaderFn,
+    TChildren
+  >(options)
+}
+
+export type AnyRootRoute = RootRoute<any, any, any, any, any, any, any>
+
+export type RootRouteOptions<
+  TSearchValidator = undefined,
+  TRouterContext = {},
+  TRouteContextFn = AnyContext,
+  TBeforeLoadFn = AnyContext,
+  TLoaderDeps extends Record<string, any> = {},
+  TLoaderFn = undefined,
+> = Omit<
+  RouteOptions<
+    any, // TParentRoute
+    RootRouteId, // TId
+    RootRouteId, // TCustomId
+    '', // TFullPath
+    '', // TPath
+    TSearchValidator,
+    {}, // TParams
+    TLoaderDeps,
+    TLoaderFn,
+    TRouterContext,
+    TRouteContextFn,
+    TBeforeLoadFn
+  >,
+  | 'path'
+  | 'id'
+  | 'getParentRoute'
+  | 'caseSensitive'
+  | 'parseParams'
+  | 'stringifyParams'
+  | 'params'
+>
+
+export function createRootRouteWithContext<TRouterContext extends {}>() {
+  return <
+    TRouteContextFn = AnyContext,
+    TBeforeLoadFn = AnyContext,
+    TSearchValidator = undefined,
+    TLoaderDeps extends Record<string, any> = {},
+    TLoaderFn = undefined,
+  >(
+    options?: RootRouteOptions<
+      TSearchValidator,
+      TRouterContext,
+      TRouteContextFn,
+      TBeforeLoadFn,
+      TLoaderDeps,
+      TLoaderFn
+    >,
+  ) => {
+    return createRootRoute<
+      TSearchValidator,
+      TRouterContext,
+      TRouteContextFn,
+      TBeforeLoadFn,
+      TLoaderDeps,
+      TLoaderFn
+    >(options as any)
+  }
+}
+
+/**
+ * @deprecated Use the `createRootRouteWithContext` function instead.
+ */
+export const rootRouteWithContext = createRootRouteWithContext
+
+export class RootRoute<
+  in out TSearchValidator = undefined,
+  in out TRouterContext = {},
+  in out TRouteContextFn = AnyContext,
+  in out TBeforeLoadFn = AnyContext,
+  in out TLoaderDeps extends Record<string, any> = {},
+  in out TLoaderFn = undefined,
+  in out TChildren = unknown,
+  in out TFileRouteTypes = unknown,
+> extends Route<
+  any, // TParentRoute
+  '/', // TPath
+  '/', // TFullPath
+  string, // TCustomId
+  RootRouteId, // TId
+  TSearchValidator, // TSearchValidator
+  {}, // TParams
+  TRouterContext,
+  TRouteContextFn,
+  TBeforeLoadFn,
+  TLoaderDeps,
+  TLoaderFn,
+  TChildren // TChildren
+> {
+  /**
+   * @deprecated `RootRoute` is now an internal implementation detail. Use `createRootRoute()` instead.
+   */
+  constructor(
+    options?: RootRouteOptions<
+      TSearchValidator,
+      TRouterContext,
+      TRouteContextFn,
+      TBeforeLoadFn,
+      TLoaderDeps,
+      TLoaderFn
+    >,
+  ) {
+    super(options as any)
+  }
+
+  addChildren<
+    const TNewChildren extends
+    | Record<string, AnyRoute>
+    | ReadonlyArray<AnyRoute>,
+  >(
+    children: TNewChildren,
+  ): RootRoute<
+    TSearchValidator,
+    TRouterContext,
+    TRouteContextFn,
+    TBeforeLoadFn,
+    TLoaderDeps,
+    TLoaderFn,
+    TNewChildren,
+    TFileRouteTypes
+  > {
+    super.addChildren(children)
+    return this as unknown as RootRoute<
+      TSearchValidator,
+      TRouterContext,
+      TRouteContextFn,
+      TBeforeLoadFn,
+      TLoaderDeps,
+      TLoaderFn,
+      TNewChildren,
+      TFileRouteTypes
+    >
+  }
+
+  _addFileChildren<const TNewChildren>(
+    children: TNewChildren,
+  ): RootRoute<
+    TSearchValidator,
+    TRouterContext,
+    TRouteContextFn,
+    TBeforeLoadFn,
+    TLoaderDeps,
+    TLoaderFn,
+    TNewChildren,
+    TFileRouteTypes
+  > {
+    super._addFileChildren(children)
+    return this as unknown as RootRoute<
+      TSearchValidator,
+      TRouterContext,
+      TRouteContextFn,
+      TBeforeLoadFn,
+      TLoaderDeps,
+      TLoaderFn,
+      TNewChildren,
+      TFileRouteTypes
+    >
+  }
+
+  _addFileTypes<TFileRouteTypes>(): RootRoute<
+    TSearchValidator,
+    TRouterContext,
+    TRouteContextFn,
+    TBeforeLoadFn,
+    TLoaderDeps,
+    TLoaderFn,
+    TChildren,
+    TFileRouteTypes
+  > {
+    return this as any
+  }
+}
+
+export function createRootRoute<
+  TSearchValidator = undefined,
+  TRouterContext = {},
+  TRouteContextFn = AnyContext,
+  TBeforeLoadFn = AnyContext,
+  TLoaderDeps extends Record<string, any> = {},
+  TLoaderFn = undefined,
+>(
+  options?: RootRouteOptions<
+    TSearchValidator,
+    TRouterContext,
+    TRouteContextFn,
+    TBeforeLoadFn,
+    TLoaderDeps,
+    TLoaderFn
+  >,
+) {
+  return new RootRoute<
+    TSearchValidator,
+    TRouterContext,
+    TRouteContextFn,
+    TBeforeLoadFn,
+    TLoaderDeps,
+    TLoaderFn
+  >(options)
+}
+
+export type ResolveFullPath<
+  TParentRoute extends AnyRoute,
+  TPath extends string,
+  TPrefixed = RoutePrefix<TParentRoute['fullPath'], TPath>,
+> = TPrefixed extends RootRouteId ? '/' : TPrefixed
+
+type RoutePrefix<
+  TPrefix extends string,
+  TPath extends string,
+> = string extends TPath
+  ? RootRouteId
+  : TPath extends string
+  ? TPrefix extends RootRouteId
+  ? TPath extends '/'
+  ? '/'
+  : `/${TrimPath<TPath>}`
+  : `${TPrefix}/${TPath}` extends '/'
+  ? '/'
+  : `/${TrimPathLeft<`${TrimPathRight<TPrefix>}/${TrimPath<TPath>}`>}`
+  : never
+
+export type TrimPath<T extends string> = '' extends T
+  ? ''
+  : TrimPathRight<TrimPathLeft<T>>
+
+export type TrimPathLeft<T extends string> =
+  T extends `${RootRouteId}/${infer U}`
+  ? TrimPathLeft<U>
+  : T extends `/${infer U}`
+  ? TrimPathLeft<U>
+  : T
+export type TrimPathRight<T extends string> = T extends '/'
+  ? '/'
+  : T extends `${infer U}/`
+  ? TrimPathRight<U>
+  : T
+
+// export type RouteMask<TRouteTree extends AnyRoute> = {
+//   routeTree: TRouteTree
+//   from: RoutePaths<TRouteTree>
+//   to?: any
+//   params?: any
+//   search?: any
+//   hash?: any
+//   state?: any
+//   unmaskOnReload?: boolean
+// }
+
+// export function createRouteMask<
+//   TRouteTree extends AnyRoute,
+//   TFrom extends string,
+//   TTo extends string,
+// >(
+//   opts: {
+//     routeTree: TRouteTree
+//   } & ToMaskOptions<Router<TRouteTree, 'never'>, TFrom, TTo>,
+// ): RouteMask<TRouteTree> {
+//   return opts as any
+// }
+
+/**
+ * @deprecated Use `ErrorComponentProps` instead.
+ */
+export type ErrorRouteProps = {
+  error: unknown
+  info?: { componentStack: string }
+  reset: () => void
 }
 
 export type ErrorComponentProps = {
@@ -14,12 +1664,15 @@ export type ErrorComponentProps = {
   info?: { componentStack: string }
   reset: () => void
 }
+export type NotFoundRouteProps = {
+  // TODO: Make sure this is `| null | undefined` (this is for global not-founds)
+  data: unknown
+}
+//
 
-export type SearchMiddleware<TSearchSchema> = (
-  ctx: SearchMiddlewareContext<TSearchSchema>,
-) => TSearchSchema
+export type SvelteNode = Component
 
-export type SyncRouteComponent<TProps> = (props: TProps) => Component
+export type SyncRouteComponent<TProps> = ((props: TProps) => SvelteNode)
 
 export type AsyncRouteComponent<TProps> = SyncRouteComponent<TProps> & {
   preload?: () => Promise<void>
@@ -28,3 +1681,60 @@ export type AsyncRouteComponent<TProps> = SyncRouteComponent<TProps> & {
 export type RouteComponent<TProps = any> = AsyncRouteComponent<TProps>
 
 export type ErrorRouteComponent = RouteComponent<ErrorComponentProps>
+
+export type NotFoundRouteComponent = SyncRouteComponent<NotFoundRouteProps>
+
+export class NotFoundRoute<
+  TParentRoute extends AnyRootRoute,
+  TRouterContext = AnyContext,
+  TRouteContextFn = AnyContext,
+  TBeforeLoadFn = AnyContext,
+  TSearchValidator = undefined,
+  TLoaderDeps extends Record<string, any> = {},
+  TLoaderFn = undefined,
+  TChildren = unknown,
+> extends Route<
+  TParentRoute,
+  '/404',
+  '/404',
+  '404',
+  '404',
+  TSearchValidator,
+  {},
+  TRouterContext,
+  TRouteContextFn,
+  TBeforeLoadFn,
+  TLoaderDeps,
+  TLoaderFn,
+  TChildren
+> {
+  constructor(
+    options: Omit<
+      RouteOptions<
+        TParentRoute,
+        string,
+        string,
+        string,
+        string,
+        TSearchValidator,
+        {},
+        TLoaderDeps,
+        TLoaderFn,
+        TRouterContext,
+        TRouteContextFn,
+        TBeforeLoadFn
+      >,
+      | 'caseSensitive'
+      | 'parseParams'
+      | 'stringifyParams'
+      | 'path'
+      | 'id'
+      | 'params'
+    >,
+  ) {
+    super({
+      ...(options as any),
+      id: '404',
+    })
+  }
+}
